+++ lib/riverpod_action_listener.dart
@@ -2,3 +2,4 @@ export 'src/result/action_result.dart';
 export 'src/result/multi_action_result.dart';
 export 'src/config/action_listener_config.dart';
 export 'src/hook/action_listener_hook.dart';
+export 'src/notifier/action_notifier.dart';