+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.5
+
+* Automatically handled via loading state; removed manual checks.
+* Unified all ActionNotifier implementations to the new base class.
+
+
 ## 0.0.4
 
 * Updated dependencies